fix(dev-server): centralize stream readiness rehydrate
