+++ packages/frontend/src/lib/browser-live/constants.ts
@@ -1,3 +1,2 @@
-export const BROWSER_LIVE_CONNECTED_EVENT_KIND = "connected";
 export const BROWSER_LIVE_RECONNECTED_EVENT_KIND = "reconnected";
 export const BROWSER_LIVE_STREAM_WARNING_EVENT_KIND = "stream-warning";
