+++ packages/frontend/src/pages/agents/right-panel/use-agent-studio-dev-server-panel.hook.test.tsx
@@ -1,13 +1,18 @@
 import { afterEach, beforeEach, describe, expect, mock, test } from "bun:test";
 import { GlobalRegistrator } from "@happy-dom/global-registrator";
-import type { DevServerGroupState, DevServerScriptState } from "@openducktor/contracts";
+import type { DevServerGroupState } from "@openducktor/contracts";
 import { useQueryClient } from "@tanstack/react-query";
 import { act, render, waitFor } from "@testing-library/react";
 import { createQueryClient } from "@/lib/query-client";
 import { QueryProvider } from "@/lib/query-provider";
 import { devServerQueryKeys } from "@/state/queries/dev-servers";
 import { restoreMockedModules } from "@/test-utils/mock-module-cleanup";
-import type { RepoSettingsInput } from "@/types/state-slices";
+import {
+  buildScript,
+  buildState,
+  createDeferred,
+  repoSettings,
+} from "./use-agent-studio-dev-server-panel-test-fixtures";
 import { renderDevServerPanelHook } from "./use-agent-studio-dev-server-panel-test-harness";
 
 const actualHostClientModule = await import("@/lib/host-client");
@@ -15,60 +20,6 @@ const actualHostClientModule = await import("@/lib/host-client");
 if (typeof document === "undefined") {
   GlobalRegistrator.register();
 }
-
-const createDeferred = <T,>() => {
-  let resolve: ((value: T | PromiseLike<T>) => void) | null = null;
-  let reject: ((reason?: unknown) => void) | null = null;
-  const promise = new Promise<T>((res, rej) => {
-    resolve = res;
-    reject = rej;
-  });
-
-  return {
-    promise,
-    resolve: (value: T) => resolve?.(value),
-    reject: (reason?: unknown) => reject?.(reason),
-  };
-};
-
-const buildScript = (overrides: Partial<DevServerScriptState> = {}): DevServerScriptState => ({
-  scriptId: "frontend",
-  name: "Frontend",
-  command: "bun run dev",
-  status: "stopped",
-  pid: null,
-  startedAt: null,
-  exitCode: null,
-  lastError: null,
-  bufferedTerminalChunks: [],
-  ...overrides,
-});
-
-const buildState = (overrides: Partial<DevServerGroupState> = {}): DevServerGroupState => ({
-  repoPath: "/repo",
-  taskId: "task-7",
-  worktreePath: "/tmp/worktree/task-7",
-  scripts: [buildScript()],
-  updatedAt: "2026-03-19T15:30:00.000Z",
-  ...overrides,
-});
-
-const repoSettings: RepoSettingsInput = {
-  defaultRuntimeKind: "opencode",
-  worktreeBasePath: "",
-  branchPrefix: "",
-  defaultTargetBranch: { remote: "origin", branch: "main" },
-  preStartHooks: [],
-  postCompleteHooks: [],
-  devServers: [{ id: "frontend", name: "Frontend", command: "bun run dev" }],
-  worktreeCopyPaths: [],
-  agentDefaults: {
-    spec: null,
-    planner: null,
-    build: null,
-    qa: null,
-  },
-};
 
 let devServerGetState = async (_repoPath: string, _taskId: string): Promise<DevServerGroupState> =>
   buildState();
@@ -121,47 +72,6 @@ afterEach(async () => {
 });
 
 describe("useAgentStudioDevServerPanel", () => {
-  test("subscribes to dev-server events while enabled so startup events are not missed", async () => {
-    const { useAgentStudioDevServerPanel } = await import("./use-agent-studio-dev-server-panel");
-    type HookArgs = Parameters<typeof useAgentStudioDevServerPanel>[0];
-    type HookResult = ReturnType<typeof useAgentStudioDevServerPanel>;
-
-    let latest: HookResult | null = null;
-    const getLatest = (): HookResult => {
-      if (latest === null) {
-        throw new Error("Hook result not ready");
-      }
-      return latest;
-    };
-
-    const Harness = ({ args }: { args: HookArgs }) => {
-      latest = useAgentStudioDevServerPanel(args);
-      return null;
-    };
-
-    const view = render(
-      <QueryProvider useIsolatedClient>
-        <Harness
-          args={{
-            repoPath: "/repo",
-            taskId: "task-7",
-            repoSettings,
-            enabled: true,
-          }}
-        />
-      </QueryProvider>,
-    );
-
-    try {
-      await waitFor(() => {
-        expect(getLatest().mode).toBe("stopped");
-      });
-      expect(devServerEventListener).not.toBeNull();
-    } finally {
-      view.unmount();
-    }
-  });
-
   test("does not query dev-server state when the repository has no configured dev servers", async () => {
     const { useAgentStudioDevServerPanel } = await import("./use-agent-studio-dev-server-panel");
     type HookArgs = Parameters<typeof useAgentStudioDevServerPanel>[0];
@@ -1697,358 +1607,6 @@ describe("useAgentStudioDevServerPanel", () => {
     } finally {
       harness.unmount();
       queryClient.clear();
-    }
-  });
-
-  test("rehydrates buffered terminal replay after a browser-live stream warning", async () => {
-    const { useAgentStudioDevServerPanel } = await import("./use-agent-studio-dev-server-panel");
-    type HookArgs = Parameters<typeof useAgentStudioDevServerPanel>[0];
-    type HookResult = ReturnType<typeof useAgentStudioDevServerPanel>;
-
-    const initialState = buildState({
-      scripts: [
-        buildScript({
-          status: "running",
-          pid: 4242,
-          startedAt: "2026-03-19T15:30:00.000Z",
-        }),
-      ],
-    });
-    const refreshedState = buildState({
-      updatedAt: "2026-03-19T15:31:00.000Z",
-      scripts: [
-        buildScript({
-          status: "running",
-          pid: 4242,
-          startedAt: "2026-03-19T15:30:00.000Z",
-          bufferedTerminalChunks: [
-            {
-              scriptId: "frontend",
-              sequence: 4,
-              data: "rehydrated output\r\n",
-              timestamp: "2026-03-19T15:31:00.000Z",
-            },
-          ],
-        }),
-      ],
-    });
-    let getStateCalls = 0;
-    devServerGetState = async () => {
-      getStateCalls += 1;
-      return getStateCalls === 1 ? initialState : refreshedState;
-    };
-    const restartDeferred = createDeferred<DevServerGroupState>();
-    devServerRestart = async () => restartDeferred.promise;
-
-    let latest: HookResult | null = null;
-    const getLatest = (): HookResult => {
-      if (latest === null) {
-        throw new Error("Hook result not ready");
-      }
-      return latest;
-    };
-
-    const Harness = ({ args }: { args: HookArgs }) => {
-      latest = useAgentStudioDevServerPanel(args);
-      return null;
-    };
-
-    const view = render(
-      <QueryProvider useIsolatedClient>
-        <Harness
-          args={{
-            repoPath: "/repo",
-            taskId: "task-7",
-            repoSettings,
-            enabled: true,
-          }}
-        />
-      </QueryProvider>,
-    );
-
-    try {
-      await waitFor(() => {
-        expect(getLatest().mode).toBe("active");
-      });
-
-      await act(async () => {
-        getLatest().onRestart();
-      });
-      await waitFor(() => {
-        expect(devServerEventListener).not.toBeNull();
-      });
-
-      await act(async () => {
-        devServerEventListener?.({
-          __openducktorBrowserLive: true,
-          kind: "stream-warning",
-          message: "Dev server stream skipped 4 events; reconnect will replay buffered events.",
-        });
-      });
-
-      await waitFor(() => {
-        expect(getLatest().selectedScriptTerminalBuffer?.entries[0]?.data).toBe(
-          "rehydrated output\r\n",
-        );
-      });
-      expect(getStateCalls).toBe(2);
-      restartDeferred.resolve(refreshedState);
-    } finally {
-      view.unmount();
-    }
-  });
-
-  test("rehydrates full buffered terminal replay when the dev-server subscription first becomes ready", async () => {
-    const { useAgentStudioDevServerPanel } = await import("./use-agent-studio-dev-server-panel");
-    type HookArgs = Parameters<typeof useAgentStudioDevServerPanel>[0];
-    type HookResult = ReturnType<typeof useAgentStudioDevServerPanel>;
-
-    const bufferedTerminalChunks = Array.from({ length: 300 }, (_, sequence) => ({
-      scriptId: "frontend",
-      sequence,
-      data: `subscription-gap output ${sequence}\r\n`,
-      timestamp: "2026-03-19T15:31:00.000Z",
-    }));
-    const initialState = buildState({
-      scripts: [
-        buildScript({
-          status: "running",
-          pid: 4242,
-          startedAt: "2026-03-19T15:30:00.000Z",
-        }),
-      ],
-    });
-    const refreshedState = buildState({
-      updatedAt: "2026-03-19T15:31:00.000Z",
-      scripts: [
-        buildScript({
-          status: "running",
-          pid: 4242,
-          startedAt: "2026-03-19T15:30:00.000Z",
-          bufferedTerminalChunks,
-        }),
-      ],
-    });
-    let getStateCalls = 0;
-    devServerGetState = async () => {
-      getStateCalls += 1;
-      return getStateCalls === 1 ? initialState : refreshedState;
-    };
-    const subscriptionReady = createDeferred<() => void>();
-    subscribeDevServerEventsMock = async (listener: (payload: unknown) => void) => {
-      devServerEventListener = listener;
-      return subscriptionReady.promise;
-    };
-
-    const harness = renderDevServerPanelHook<HookArgs, HookResult>(useAgentStudioDevServerPanel, {
-      repoPath: "/repo",
-      taskId: "task-7",
-      repoSettings,
-      enabled: true,
-    });
-
-    try {
-      await waitFor(() => {
-        expect(harness.getLatest().mode).toBe("active");
-      });
-      expect(getStateCalls).toBe(1);
-
-      subscriptionReady.resolve(() => {
-        devServerEventListener = null;
-      });
-
-      await waitFor(() => {
-        expect(harness.getLatest().selectedScriptTerminalBuffer?.entries).toHaveLength(300);
-      });
-      expect(harness.getLatest().selectedScriptTerminalBuffer?.entries[0]?.data).toBe(
-        "subscription-gap output 0\r\n",
-      );
-      expect(harness.getLatest().selectedScriptTerminalBuffer?.entries.at(-1)?.data).toBe(
-        "subscription-gap output 299\r\n",
-      );
-      expect(getStateCalls).toBe(2);
-    } finally {
-      harness.unmount();
-    }
-  });
-
-  test("rehydrates buffered terminal replay after a browser-live reconnect", async () => {
-    const { useAgentStudioDevServerPanel } = await import("./use-agent-studio-dev-server-panel");
-    type HookArgs = Parameters<typeof useAgentStudioDevServerPanel>[0];
-    type HookResult = ReturnType<typeof useAgentStudioDevServerPanel>;
-
-    const initialState = buildState({
-      scripts: [
-        buildScript({
-          status: "running",
-          pid: 4242,
-          startedAt: "2026-03-19T15:30:00.000Z",
-          bufferedTerminalChunks: [
-            {
-              scriptId: "frontend",
-              sequence: 0,
-              data: "stale output\r\n",
-              timestamp: "2026-03-19T15:30:00.000Z",
-            },
-          ],
-        }),
-      ],
-    });
-    const refreshedState = buildState({
-      updatedAt: "2026-03-19T15:31:00.000Z",
-      scripts: [
-        buildScript({
-          status: "running",
-          pid: 4242,
-          startedAt: "2026-03-19T15:30:00.000Z",
-          bufferedTerminalChunks: [
-            {
-              scriptId: "frontend",
-              sequence: 2,
-              data: "reconnected output\r\n",
-              timestamp: "2026-03-19T15:31:00.000Z",
-            },
-          ],
-        }),
-      ],
-    });
-    let getStateCalls = 0;
-    devServerGetState = async () => {
-      getStateCalls += 1;
-      return getStateCalls === 1 ? initialState : refreshedState;
-    };
-
-    let latest: HookResult | null = null;
-    const getLatest = (): HookResult => {
-      if (latest === null) {
-        throw new Error("Hook result not ready");
-      }
-      return latest;
-    };
-
-    const Harness = ({ args }: { args: HookArgs }) => {
-      latest = useAgentStudioDevServerPanel(args);
-      return null;
-    };
-
-    const view = render(
-      <QueryProvider useIsolatedClient>
-        <Harness
-          args={{
-            repoPath: "/repo",
-            taskId: "task-7",
-            repoSettings,
-            enabled: true,
-          }}
-        />
-      </QueryProvider>,
-    );
-
-    try {
-      await waitFor(() => {
-        expect(getLatest().selectedScriptTerminalBuffer?.entries[0]?.data).toBe("stale output\r\n");
-      });
-      await waitFor(() => {
-        expect(devServerEventListener).not.toBeNull();
-      });
-
-      await act(async () => {
-        devServerEventListener?.({
-          __openducktorBrowserLive: true,
-          kind: "reconnected",
-        });
-      });
-
-      await waitFor(() => {
-        expect(getLatest().selectedScriptTerminalBuffer?.entries[0]?.data).toBe(
-          "reconnected output\r\n",
-        );
-      });
-      expect(getStateCalls).toBe(2);
-    } finally {
-      view.unmount();
-    }
-  });
-
-  test("surfaces invalid dev server event payloads as actionable errors", async () => {
-    const { useAgentStudioDevServerPanel } = await import("./use-agent-studio-dev-server-panel");
-    type HookArgs = Parameters<typeof useAgentStudioDevServerPanel>[0];
-    type HookResult = ReturnType<typeof useAgentStudioDevServerPanel>;
-
-    devServerGetState = async () =>
-      buildState({
-        scripts: [
-          buildScript({
-            status: "running",
-            pid: 4242,
-            startedAt: "2026-03-19T15:30:00.000Z",
-          }),
-        ],
-      });
-    const restartDeferred = createDeferred<DevServerGroupState>();
-    devServerRestart = async () => restartDeferred.promise;
-
-    let latest: HookResult | null = null;
-    const getLatest = (): HookResult => {
-      if (latest === null) {
-        throw new Error("Hook result not ready");
-      }
-      return latest;
-    };
-
-    const Harness = ({ args }: { args: HookArgs }) => {
-      latest = useAgentStudioDevServerPanel(args);
-      return null;
-    };
-
-    const view = render(
-      <QueryProvider useIsolatedClient>
-        <Harness
-          args={{
-            repoPath: "/repo",
-            taskId: "task-7",
-            repoSettings,
-            enabled: true,
-          }}
-        />
-      </QueryProvider>,
-    );
-    const originalConsoleError = console.error;
-    console.error = () => {};
-
-    try {
-      await waitFor(() => {
-        expect(getLatest().mode).toBe("active");
-      });
-
-      await act(async () => {
-        getLatest().onRestart();
-      });
-      await waitFor(() => {
-        expect(devServerEventListener).not.toBeNull();
-      });
-
-      act(() => {
-        devServerEventListener?.({ type: "terminal_chunk", repoPath: "/repo", taskId: "task-7" });
-      });
-
-      await waitFor(() => {
-        expect(getLatest().error).toContain("Received invalid dev server event payload");
-      });
-      restartDeferred.resolve(
-        buildState({
-          scripts: [
-            buildScript({
-              status: "running",
-              pid: 4242,
-              startedAt: "2026-03-19T15:30:00.000Z",
-            }),
-          ],
-        }),
-      );
-    } finally {
-      console.error = originalConsoleError;
-      view.unmount();
     }
   });
